refactor to remove boilerplate codes
